refactor(代码优化): 优化订单VO价格字段类型，修复注册bug

+++ src/main/java/ltd/newbee/mall/controller/vo/OrderDetailVO.java
@@ -10,7 +10,7 @@ public class OrderDetailVO {
 
     private String orderNo;
 
-    private Double totalPrice;
+    private Integer totalPrice;
 
     private Byte payStatus;
